# functions/swagger.yaml
openapi: 3.0.3
info:
  title: AiDuxCare Patient API
  description: RESTful API for managing patients in AiDuxCare backend.
  version: 1.0.0
servers:
  - url: https://YOUR_API_GATEWAY_DOMAIN
paths:
  /patients:
    get:
      summary: List all patients
      operationId: listPatients
      responses:
        '200':
          description: List of patients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Patient'
    post:
      summary: Create a new patient
      operationId: createPatient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientDTO'
      responses:
        '201':
          description: Patient created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '400':
          description: Invalid input
  /patients/{id}:
    get:
      summary: Get a patient by ID
      operationId: getPatient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Patient found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '404':
          description: Patient not found
    put:
      summary: Update a patient by ID
      operationId: updatePatient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePatientDTO'
      responses:
        '200':
          description: Patient updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Patient'
        '404':
          description: Patient not found
    delete:
      summary: Delete a patient by ID
      operationId: deletePatient
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Patient deleted
        '404':
          description: Patient not found
components:
  schemas:
    Patient:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        birthDate:
          type: string
          format: date
        reasonForConsultation:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    CreatePatientDTO:
      type: object
      required:
        - name
        - email
        - phone
        - birthDate
        - reasonForConsultation
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        birthDate:
          type: string
          format: date
        reasonForConsultation:
          type: string 